chore: reorganise the code
