+++ src/app/landing/page.tsx
@@ -1,0 +1,5 @@
+import DashboardPage from "./@expenses/page";
+
+export default function Page() {
+  return <DashboardPage />;
+}
