+++ src/app/page.tsx
@@ -1,9 +1,5 @@
-import DashboardPage from "./dashboard/page";
+import { redirect } from 'next/navigation';
 
 export default function Home() {
-  return (
-      <div>
-        something
-      </div>
-  );
+  redirect('/landing');
 }
